More alpha experiment results.

+++ tests/race_checker/trace_linux.c
@@ -17,11 +17,11 @@ void trace_attach(pid_t pid) {
 }
 
 void trace_step(pid_t pid, int signum) {
-    error_wrap(ptrace(PTRACE_SINGLESTEP, pid, NULL, (void*) signum),
+    error_wrap(ptrace(PTRACE_SINGLESTEP, pid, NULL, (void*) (long) signum),
                "ptrace(PTRACE_SINGLESTEP, ...)", ERRNO);
 }
 
 void trace_detach(pid_t pid, int signum) {
-    error_wrap(ptrace(PTRACE_DETACH, pid, NULL, (void*) signum),
+    error_wrap(ptrace(PTRACE_DETACH, pid, NULL, (void*) (long) signum),
                "ptrace(PTRACE_DETACH, ...)", ERRNO);
 }
